voice-agent/requirements.txt
livekit-agents
livekit-plugins-elevenlabs
livekit-plugins-deepgram
livekit-plugins-google
livekit-plugins-silero
python-dotenv
requests
flask
flask-cors
twilio
motor
